backend/requirements_utf8.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.0
arabic-reshaper==3.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asn1crypto==1.5.1
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
bcrypt==3.2.2
beautifulsoup4==4.14.3
bleach==6.3.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cross-web==0.4.1
cryptography==46.0.5
cssselect2==0.9.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==6.0.3
docstring_parser==0.17.0
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.61.0
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.49.0
graphql-core==3.2.8
greenlet==3.3.2
groq==1.1.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kagglehub==0.4.1
kagglesdk==0.1.15
kiwisolver==1.4.9
lark==1.3.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.3.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pillow==12.0.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.1.3
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
PyJWT==2.11.0
pyparsing==3.2.5
pypdf==6.9.1
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rlPyCairo==0.4.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
strawberry-graphql==0.311.0
svglib==1.6.0
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.1.0
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritools==6.0.1
urllib3==2.5.0
uvicorn==0.41.0
wcwidth==0.2.14
weasyprint==68.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wfdb==4.3.0
widgetsnbextension==4.0.15
xhtml2pdf==0.2.17
yarl==1.22.0
zopfli==0.4.1
